+++ scout/server/utils.py
@@ -244,15 +244,23 @@ def case_has_chanjo2_coverage(case_obj: dict):
 def case_has_alignments(case_obj: dict):
     """Add info on bam/cram files availability to a case dictionary
 
+    This sets the availability of alignments for autosomal chromosomes.
+    If paraphase or de novo assembly alignments, this is also sufficient to set
+    alignment availability.
     Args:
         case_obj(scout.models.Case)
     """
-    case_obj["bam_files"] = False  # Availability of alignments for autosomal chromosomes
+    case_obj["bam_files"] = False
     for ind in case_obj.get("individuals"):
-        bam_path = ind.get("bam_file")
-        if bam_path and os.path.exists(bam_path):
-            case_obj["bam_files"] = True
-            return
+        for alignment_path_key in [
+            "bam_file",
+            "assembly_alignment_path",
+            "paraphase_alignment_path",
+        ]:
+            bam_path = ind.get(alignment_path_key)
+            if bam_path and os.path.exists(bam_path):
+                case_obj["bam_files"] = True
+                return
 
 
 def case_has_mt_alignments(case_obj: dict):
@@ -304,7 +312,9 @@ def case_append_alignments(case_obj: dict):
     """
     unwrap_settings = [
         {"path": "bam_file", "append_to": "bam_files", "index": "bai_files"},
+        {"path": "assembly_alignment_path", "append_to": "bam_files", "index": "bai_files"},
         {"path": "mt_bam", "append_to": "mt_bams", "index": "mt_bais"},
+        {"path": "paraphase_alignment_path", "append_to": "bam_files", "index": "bai_files"},
         {"path": "rhocall_bed", "append_to": "rhocall_beds", "index": None},
         {"path": "rhocall_wig", "append_to": "rhocall_wigs", "index": None},
         {"path": "upd_regions_bed", "append_to": "upd_regions_beds", "index": None},